WIP: implement watcher actors registration
