+++ .gitignore
@@ -9,6 +9,6 @@ composer.lock
 index.php
 .idea
 config.yaml
-commands/Watch/WatcherActors/Karma/store/*.json
+commands/Karma/WatcherActor/store/*.json
 tests/Fixtures/karma.tmp.json
 tests/Fixtures/karma.create-store.tmp.json
