+++ commands/Karma/WatcherActor/Exceptions/InvalidUserNameException.php
@@ -1,0 +1,7 @@
+<?php
+
+namespace SoerBot\Commands\Karma\WatcherActor\Exceptions;
+
+class InvalidUserNameException extends \Exception
+{
+}
