+++ commands/Karma/WatcherActor/Exceptions/StoreFileNotFoundException.php
@@ -1,0 +1,7 @@
+<?php
+
+namespace SoerBot\Commands\Karma\WatcherActor\Exceptions;
+
+class StoreFileNotFoundException extends \Exception
+{
+}
