--- commands/Watch/WatcherActors/Karma/Exceptions/InvalidUserNameException.php
@@ -1,7 +1,0 @@
-<?php
-
-namespace SoerBot\Commands\Watch\WatcherActors\Karma\Exceptions;
-
-class InvalidUserNameException extends \Exception
-{
-}
