--- commands/Watch/WatcherActors/Karma/Exceptions/StoreFileNotFoundException.php
@@ -1,7 +1,0 @@
-<?php
-
-namespace SoerBot\Commands\Watch\WatcherActors\Karma\Exceptions;
-
-class StoreFileNotFoundException extends \Exception
-{
-}
